+++ Assets/Scripts/Interactable/Interactable.cs
@@ -2,12 +2,7 @@ using UnityEngine;
 
 public abstract class Interactable : MonoBehaviour {
 
-    protected void Start() {
-
-        if (!transform.CompareTag("Interactable"))
-            Debug.LogWarning("Interactable object " + name + " is not tagged as Interactable! Interacting will not work for this object");
-
-    }
+    protected void Start() => transform.tag = "Interactable"; // ensure the object is tagged as Interactable
 
     public abstract void Interact();