+++ Assets/Scripts/Inventory/RepairInventory.cs
@@ -16,21 +16,30 @@ public class RepairInventory : Inventory {
         this.repairPercent = repairPercent; // set the repair percent to the percent durability of the system that is to be repaired
         this.systemType = systemType; // set the system type to the type of system that is to be repaired
 
-        // set the item whitelist to the repair stack items
-        this.itemWhitelist = new Item[repairStacks.Length];
+        this.itemTypeFilterType = FilterType.Whitelist; // set the item type filter type to whitelist since we only want to allow the item types that are in the repair stacks
+
+        this.filteredItemTypes = new ItemType[repairStacks.Length]; // set the filtered item types to the item types of the repair stacks
 
         for (int i = 0; i < repairStacks.Length; i++)
-            this.itemWhitelist[i] = repairStacks[i].GetItem();
+            this.filteredItemTypes[i] = repairStacks[i].GetItem().GetItemType(); // set the filtered item types to the item types of the repair stacks
+
+        this.itemFilterType = FilterType.Whitelist; // set the filter type to whitelist since we only want to allow the items that are in the repair stacks
+
+        // set the item whitelist to the repair stack items
+        this.filteredItems = new Item[repairStacks.Length];
+
+        for (int i = 0; i < repairStacks.Length; i++)
+            this.filteredItems[i] = repairStacks[i].GetItem();
 
         base.Initialize(); // initialize at the end to ensure the properties are set before calling the base method (especially the slot count)
 
     }
 
-    private void OnEnable() => onItemStackAdded += OnItemStackAdded;
+    private void OnEnable() => onContentsUpdate += OnItemStackAdded;
 
     private void Start() => systemRepairMenu = FindFirstObjectByType<SystemRepairMenu>();
 
-    private void OnDisable() => onItemStackAdded -= OnItemStackAdded;
+    private void OnDisable() => onContentsUpdate -= OnItemStackAdded;
 
     private void OnItemStackAdded() {
 
@@ -39,11 +48,10 @@ public class RepairInventory : Inventory {
                 return;
 
         // if we reach here, all required stacks are present for repairing
-        systemRepairMenu.OnRepairInventoryFull(repairPercent, systemType); // notify the system repair menu that the repair inventory is full, which means the necessary items for repairing were put in
+        systemRepairMenu.OnRepairRequirementsMet(repairPercent, systemType); // notify the system repair menu that the repair inventory is full, which means the necessary items for repairing were put in
 
     }
 
-    // helper to get the effective stack limit for an item in a slot
     public override int GetEffectiveStackLimit(Item item) {
 
         // return the amount of items required to repair the system if the item is in the repair stack
@@ -54,9 +62,6 @@ public class RepairInventory : Inventory {
         return 0;
 
     }
-
-    public ItemStack[] GetRepairStacks() => repairStacks;
-
 }
 
 #if UNITY_EDITOR
@@ -71,7 +76,10 @@ public class RepairInventoryEditor : Editor {
         DrawPropertiesExcluding(serializedObject,
             "initialSlotCount",
             "slotStackLimit",
-            "itemWhitelist"
+            "itemTypeFilterType",
+            "filteredItemTypes",
+            "itemFilterType",
+            "filteredItems"
         );
 
         serializedObject.ApplyModifiedProperties();