+++ Packages/com.arongranberg.astar/ExampleScenes~/ExampleScripts/DocumentationButton.cs
@@ -8,9 +8,15 @@ namespace Pathfinding.Examples {
 
 		const string UrlBase = "https://arongranberg.com/astar/docs/";
 
+		GUIContent buttonContent = new GUIContent("Example Scene Documentation");
+
+		void Awake () {
+			useGUILayout = false;
+		}
+
 #if UNITY_EDITOR
 		void OnGUI () {
-			if (GUI.Button(new Rect(Screen.width - 250, Screen.height - 60, 240, 50), "Example Scene Documentation")) {
+			if (GUI.Button(new Rect(Screen.width - 250, Screen.height - 60, 240, 50), buttonContent)) {
 				Application.OpenURL(UrlBase + page + ".html");
 			}
 		}