+++ Packages/com.arongranberg.astar/Graphs/RecastGraph.cs
@@ -1500,7 +1500,6 @@ namespace Pathfinding {
 			}
 			EndBatchTileUpdate();
 
-			// TODO: Call after ReplaceTile too? Make sure it's not called after navmesh cut updates
 			if (OnRecalculatedTiles != null) OnRecalculatedTiles(updatedTiles);
 		}