+++ Packages/com.arongranberg.astar/Utilities/JobDependencyTracker.cs
@@ -34,21 +34,22 @@ namespace Pathfinding.Jobs {
 
 		public void Add<T>(NativeArray<T> data) where T : unmanaged {
 			if (buffer == null) buffer = ListPool<NativeArray<byte> >.Claim();
-			buffer.Add(data.Reinterpret<byte>(UnsafeUtility.SizeOf<T>()));
+			// SAFETY: This is safe because NativeArray<byte> and NativeArray<T> have the same memory layout.
+			// Note: The resulting array will have the wrong length, but the length is not used when disposing the array.
+			// Note: It's important *not* to use the Reinterpret method, as for large arrays with large structs, the length in bytes could overflow 32-bits.
+			buffer.Add(UnsafeUtility.As<NativeArray<T>, NativeArray<byte> >(ref data));
 		}
 
 		public void Add<T>(NativeList<T> data) where T : unmanaged {
 			// SAFETY: This is safe because NativeList<byte> and NativeList<T> have the same memory layout.
-			var byteList = Unity.Collections.LowLevel.Unsafe.UnsafeUtility.As<NativeList<T>, NativeList<byte> >(ref data);
 			if (buffer2 == null) buffer2 = ListPool<NativeList<byte> >.Claim();
-			buffer2.Add(byteList);
+			buffer2.Add(UnsafeUtility.As<NativeList<T>, NativeList<byte> >(ref data));
 		}
 
 		public void Add<T>(NativeQueue<T> data) where T : unmanaged {
 			// SAFETY: This is safe because NativeQueue<byte> and NativeQueue<T> have the same memory layout.
-			var byteList = Unity.Collections.LowLevel.Unsafe.UnsafeUtility.As<NativeQueue<T>, NativeQueue<byte> >(ref data);
 			if (buffer3 == null) buffer3 = ListPool<NativeQueue<byte> >.Claim();
-			buffer3.Add(byteList);
+			buffer3.Add(UnsafeUtility.As<NativeQueue<T>, NativeQueue<byte> >(ref data));
 		}
 
 		public void Remove<T>(NativeArray<T> data) where T : unmanaged {